# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/panels/access/{panelKey}": {
      "patch": {
        "operationId": "PanelsController_setAccess",
        "parameters": [
          {
            "name": "panelKey",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePanelAccessDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Toggle a sibling panel on/off",
        "tags": [
          "panels"
        ]
      }
    },
...
    "/it/services/internal/{key}": {
      "patch": {
        "operationId": "ItServicesController_toggleInternal",
        "parameters": [
          {
            "name": "key",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToggleInternalServiceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "روشن/خاموش کردن سرویس داخلی",
        "tags": [
          "it-manager"
        ]
      }
    },
    "/it/services/external": {
      "post": {
        "operationId": "ItServicesController_createExternal",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExternalServiceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "تعریف سرویس خارجی جدید",
        "tags": [
          "it-manager"
        ]
      }
    },
    "/it/services/external/{id}": {
      "patch": {
        "operationId": "ItServicesController_updateExternal",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateExternalServiceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "به‌روزرسانی سرویس خارجی",
        "tags": [
          "it-manager"
        ]
      },
      "delete": {
        "operationId": "ItServicesController_removeExternal",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "حذف سرویس خارجی",
        "tags": [
          "it-manager"
        ]
      }
    },
    "/it/services/external/{id}/test": {
      "post": {
        "operationId": "ItServicesController_testExternal",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "تست اتصال واقعی سرویس خارجی",
        "tags": [
          "it-manager"
        ]
      }
    },
...
    "/refunds/{id}": {
      "get": {
        "operationId": "RefundsController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "جزئیات درخواست — تنها سطحی که شبا/PII رمزگشایی‌شده می‌گیرد",
        "tags": [
          "refunds"
        ]
      }
    },
...
    "/my/refunds": {
      "post": {
        "operationId": "RefundsCustomerController_submit",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitRefundDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "ثبت درخواست استرداد بلیط با ریز جریمه/مبلغ قابل استرداد",
        "tags": [
          "refunds"
        ]
      },
      "get": {
        "operationId": "RefundsCustomerController_listMine",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "فهرست درخواست‌های استرداد مشتری جاری",
        "tags": [
          "refunds"
        ]
      }
    },
...
    "/my/refunds/preview": {
      "post": {
        "operationId": "RefundsCustomerController_preview",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewRefundDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "بازمحاسبه مالکیت، امکان استرداد و جریمه پیش از تأیید",
        "tags": [
          "refunds"
        ]
      }
    },
    "/my/refunds/{id}": {
      "get": {
        "operationId": "RefundsCustomerController_getMine",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "جزئیات یک درخواست استرداد مشتری",
        "tags": [
          "refunds"
        ]
      }
    },
...
  },
  "info": {
    "title": "blujet API",
    "description": "blujet airline platform — internal API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdatePanelAccessDto": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "enabled"
        ]
      },
...
      "ToggleInternalServiceDto": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "enabled"
        ]
      },
      "CreateExternalServiceDto": {
        "type": "object",
        "properties": {
          "nameFa": {
            "type": "string",
            "example": "درگاه پرداخت زرین‌پال"
          },
          "provider": {
            "type": "string",
            "example": "زرین‌پال"
          },
          "endpoint": {
            "type": "string",
            "example": "https://api.provider.com/v1/"
          },
          "method": {
            "type": "string",
            "enum": [
              "GET",
              "POST"
            ],
            "default": "POST"
          },
          "timeoutMs": {
            "type": "number",
            "example": 30000
          },
          "apiKey": {
            "type": "string"
          },
          "sandbox": {
            "type": "boolean",
            "default": false
          }
        },
        "required": [
          "nameFa",
          "provider",
          "endpoint"
        ]
      },
      "UpdateExternalServiceDto": {
        "type": "object",
        "properties": {
          "nameFa": {
            "type": "string"
          },
          "endpoint": {
            "type": "string"
          },
          "method": {
            "type": "string",
            "enum": [
              "GET",
              "POST"
            ]
          },
          "timeoutMs": {
            "type": "number"
          },
          "apiKey": {
            "type": "string"
          },
          "sandbox": {
            "type": "boolean"
          },
          "enabled": {
            "type": "boolean"
          }
        }
      },
...
      "SubmitRefundDto": {
        "type": "object",
        "properties": {
          "bookingId": {
            "type": "string",
            "description": "شناسه رزرو متعلق به مشتری جاری",
            "example": "c2c3954d-b315-40d2-a663-bad2d310bec2"
          },
          "iban": {
            "type": "string",
            "example": "IR820170000000332211009900",
            "description": "۲۴ رقم شبا"
          }
        },
        "required": [
          "bookingId",
          "iban"
        ]
      },
      "PreviewRefundDto": {
        "type": "object",
        "properties": {
          "bookingId": {
            "type": "string",
            "description": "شناسه رزرو متعلق به مشتری جاری",
            "example": "c2c3954d-b315-40d2-a663-bad2d310bec2"
          }
        },
        "required": [
          "bookingId"
        ]
      },
...
    }
  }
}